pyproject.toml
[tool.poetry]
name = "magicparse"
version = "1.0.2"
description = "Declarative parser 🛸"
authors = ["Smartway <dev@zero-gachis.com>"]
readme = "README.md"
repository = "https://github.com/ZeroGachis/magicparse"


[tool.poetry.dependencies]
python = "^3.13"
jsonata-python = "^0.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9"
awscli = "~1"
ruff = "^0.15"
pre-commit = "^4"
basedpyright = "^1.38"

[tool.poetry.group.ci]
optional = true

[tool.poetry.group.ci.dependencies]
pytest-cov = "^7.0.0"
pytest-github-actions-annotate-failures = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.masonry.api"


[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
ignore = [
    "E722", # Allow for bare 'except'
]

[tool.pytest.ini_options]
python_files = ["tests/*"]


[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"
include = ["magicparse/**", "tests/**"]

executionEnvironments = [{ root = "magicparse" }]
failOnWarnings = true
deprecateTypingAliases = true
